remove(ng): extract sessionsDownloader
