+++ app/javascript/angular/code/services/fixed_sessions.js
@@ -2,9 +2,4 @@ import { fixedSessions } from "./_fixed_sessions";
 
 angular
   .module("aircasting")
-  .factory("fixedSessions", [
-    "$rootScope",
-    "sessionsDownloader",
-    "$window",
-    fixedSessions
-  ]);
+  .factory("fixedSessions", ["$rootScope", "$window", fixedSessions]);
